rockchip: enable timer7 on rk3288 platforms for broken bootloaders
timer7 supplies the architected timer and thus as has to run when
the system clocksource and clockevents drivers are registered.

While it should be the responsibility of the bootloader to do this,
older uboot variants for the rk3288 did not do this.

Therefore add a snippet for those, enabling the timer early in the
kernel.